Seed isotopes
